move read to sep module
